backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.118.0",
    "openai-agents>=0.3.3",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.7.1",
    "qdrant-client>=1.15.1",
    "uvicorn[standard]>=0.37.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.1.1",
    "ruff>=0.13.3",
]
